requirements.txt
Flask
numpy
opencv-python
pillow
